updated readme to address an issue
Some students were encountering an issue with how the green_onion gem installed gem files (specifically fileutils, which is not going to be a gem in ruby versions <2.5 and remains a part of the stdlib)
Spoke with members of curriculum development and the dev team. In light of no obvious way to address the issue, 
users can force install the gem if they run into the issue. The error this suppresses (which prevents the unneeded 'fileutils' gem from installing) is inconsequential because 'fileutils' is correctly required in the main spec file regardless. 
The best solution would be for the author of green_onion to update their package.